+++ drawio/lib/Migration/MimeTypeMigration.php
@@ -1,0 +1,29 @@
+<?php
+
+/**
+ *
+ * @author Pawel Rojek <pawel at pawelrojek.com>
+ *
+ * This file is licensed under the Affero General Public License version 3 or later.
+ *
+ * Based on Keeweb solution
+ *
+ **/
+
+namespace OCA\Drawio\Migration;
+
+use OCP\Files\IMimeTypeLoader;
+use OCP\Migration\IOutput;
+use OCP\Migration\IRepairStep;
+
+abstract class MimeTypeMigration implements IRepairStep
+{
+    const CUSTOM_MIMETYPEMAPPING = 'mimetypemapping.json';
+
+    protected $mimeTypeLoader;
+
+    public function __construct(IMimeTypeLoader $mimeTypeLoader)
+    {
+        $this->mimeTypeLoader = $mimeTypeLoader;
+    }
+}